requirements.txt
altgraph==0.17
async-generator==1.10
attrs==21.4.0
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
click==8.0.1
cryptography==36.0.1
dash==2.2.0
dash-bootstrap-components==0.13.0
dash-core-components==2.0.0
dash-daq==0.5.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
et-xmlfile==1.1.0
fake-useragent==0.1.11
Flask==2.0.1
Flask-Compress==1.10.1
future==0.18.2
h11==0.13.0
idna==2.10
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
keyring==23.0.1
macholib==1.14
MarkupSafe==2.0.1
modulegraph==0.18
names==0.3.0
numpy==1.21.0
openpyxl==3.0.7
outcome==1.1.0
pandas==1.2.5
plotly==5.1.0
py2app==0.24
pycparser==2.20
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pyOpenSSL==22.0.0
pyoxidizer==0.17.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tenacity==7.0.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.6
Werkzeug==2.0.1
wsproto==1.0.0
zipp==3.5.0
